Remove new DefaultNullMarked test in nullness directory.
@NullMarked is tested by other tests. Let's avoid needing a JSpecify
dependency for the main tests.

This should fix the CI failure I'm seeing for NullnessTest. But it won't
fix the CI failure I'm seeing for NullnessNullMarkedSubpackageTest. I
wonder if the JSpecify jar isn't being built by the CI scripts? That
would have been fine when the jar was needed only for
NullnessJSpecifySamples, which isn't run by default, but it's a problem
now. I'll investigate.