Improve the suggested Maven configuration (#6121)
This change includes multiple adjustments:

- Instead of adding `com.google.errorprone:javac` as a dependency, copy
it to the output folder. This is important to ensure that your project
does not accidentally import and use any of the classes in that jar. It
will also reduce the size of your project's dependency tree.
Furthermore, this is only relevant in the JDK 8 profile.

  - Removed unused property `errorProneJavac`.

  - Removed (now-)unused use of `maven-dependency-plugin:properties`.

  - Converted deprecated `<compilerArguments>` to `<compilerArgs>`.

- Removed explicit dependency on `org.checkerframework:checker`, which
is not necessary because it is also present in the
`<annotationProcessorPaths>` block. This will also reduce the size of
your project's dependency tree (you probably only need the classes in
`checker-qual`).

- Removed unnecessary `<fork>true</fork>` declarations; these are
inherited.

- Moved animal-sniffer configuration to its own section. I don't think
we should recommend a blanket disable; people might copy-paste the
configuration without realizing they are disabling a useful plugin.

- Added single-quotes in the `-P '!checkerframework'` flags. The bang
character has special meaning in most shells (including Bash), and the
quotes are almost certainly required for anyone copy-pasting this code
into their shell.

- Updated the configuration in .../MavenExample/pom.xml to align with
what's in the manual. I confirmed that the configuration works with
OpenJDK 8 and 20.