Provide a method of declaring the location of test source fixtures; fixes #6125 (#6132)
The root directory for test fixtures is required in two places:

- in the test class so it can be configured with the correct source path by org.checkerframework.framework.test.TestConfigurationBuilder to compile the fixtures.
- in the test runner so it can find the test fixtures and inject them as the test's parameters.

Introduce a `@TestRootDirectory` annotation for tests to declare the location their test source fixtures. The absence of the annotation is the status-quo and the location will default to the current behaviour.

Refactoring of `CheckerFrameworkPerDirectoryTest` and `CheckerFrameworkPerFileTest` introduced a common superclass that may also be suitable for addressing the extant to do notes in these classes.