Speed up Spotless formatting (#6383)
Fixes #6297

We make two key changes:

1. Split formatting of `.java` source files among individual sub-projects, so formatting / checking for each sub-project can run in parallel.
2. Remove exclusion patterns that start in `**`, which are expensive to evaluate.

With these changes, `./gradlew spotlessCheck` (which runs in the pre-commit script) runs much faster for me (on my M1 Mac the running time is reduced from ~9.2 seconds before this PR to ~2 seconds after).

To test, we add tasks named `printSpotlessTaskInputs` that print which files are being formatted.  Using these tasks I confirmed that exactly the same files are being checked and formatted before and after this PR.